Add a lot of documentation around the (updated) CE scripts: sts/nli/quora/rerankers/ms_marco/distillation
